+++ docs/changes.rst
@@ -4,8 +4,9 @@ Change Log
 
 * 0.2.8 (dev)
 
-  * created :func:`pysftp.walktree`
-
+  * created :func:`pysftp.walktree` for walking local directories
+  * added param recurse to :meth:`.pysftp.Connection.walktree` to allow it to do another trick
+  * created :meth:`.put_d` to put the contents of a local directory to a remote one
 
 * 0.2.7 (current, released 2014-05-24)